feat: some fixes
