--- p2p_wallet/Scenes/Main/Swap/Common/Models/JupiterSwapError.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-protocol JupiterSwapError: Error {}
-
-enum JupiterSwapGeneralError: JupiterSwapError {
-    case networkError
-    case unknown
-}
